refactor: extract reconcileConfirmedPrompt into dedicated module
